Add variable to configure servers and options
It is currently not possible to override the servers without
overriding the entire configuration which can result in being
different from the repo configuration.

This change introduces two new variables that can be used for
configuration:

- ntp_servers
- ntp_server_options

Their default values match the same ones that exist already
in the configuration, but it will now allow modifying the
server options and use the existing default servers, or
changing all the server config all together.